finalize bounties/notify-partners cron
